+++ RELEASE_NOTES.md
@@ -4,6 +4,10 @@
 * Fix bug of exchange wrongly tagged as biosphere when imported from excel
 * Add setting to prevent overriding parameters
 
+# 1.3.1 
+
+* Improved speed of symbolic matrix inversion using sparse matrices and "LU" method for inverse
+
 # 1.3
 
 * Added support for "loops" in fireground inventory : 
@@ -11,6 +15,13 @@
   Internally, we now build a dict of <BG-Act, Expression> rarther than a single big expression that used custom 
   symbols as BG activities. This makes the code much cleaner and may tackle some performance issues with big inventories.
 * Added support of brightway 25 is a separate branch/package lca_algebraic_bw25
+
+
+# 1.2.2
+
+* Fix bug #78 : setOutputAmount() duplicates output exchange.
+* Backported fix for bug #66 : Fix freezeParam to use user defined functions
+* Backported #76 : Fix and improve copyActivity(withExchange=False)
 
 # 1.2